+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * fixed stool create gav in shared mode (thanks to Maximilian Braun)
 * Dashboard: removed cancel button, it had no effect (thanks to Cosmin Barbu)
+* one logstash process per stage
 
 
 ### 3.4.2 (2016-08-11)